# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/media/storage-status": {
      "get": {
        "operationId": "media_storageStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Whether uploads can be accepted and how many clips the caller has left in\ntheir plan's window, so the UI can say so before recording.",
        "tags": ["media"]
      }
    },
    "/api/v1/media/upload-url": {
      "post": {
        "operationId": "media_requestUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["media"]
      }
    },
    "/api/v1/media/confirm": {
      "post": {
        "operationId": "media_confirmUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["media"]
      }
    },
    "/api/v1/media/recent": {
      "get": {
        "operationId": "media_listRecent",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The newest clips platform-wide, each with its player. Drives the landing\nstrip in one request instead of one per player — see MediaService.listRecent.\n\nDeclared before `player/:playerId` and `:id`, since Nest matches in\ndeclaration order.",
        "tags": ["media"]
      }
    },
    "/api/v1/media/feed": {
      "get": {
        "operationId": "media_feed",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 24,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The ranked feed — the scout's and the academy manager's home screen.\n\nSigned in only: the ranking is personalised by who the caller follows and\nmarks what they have already liked, neither of which exists for a guest.\nDeclared before `player/:playerId` and `:id`, since Nest matches in\ndeclaration order.",
        "tags": ["media"]
      }
    },
    "/api/v1/media/feed/suggested-players": {
      "get": {
        "operationId": "media_suggestedPlayers",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Players worth following, for the panel beside the feed.",
        "tags": ["media"]
      }
    },
    "/api/v1/media/player/{playerId}": {
      "get": {
        "operationId": "media_listForPlayer",
        "parameters": [
          {
            "name": "playerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/MediaCategory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "A player's clips, newest first. `category` filters to one attribute's history.\n\nCarries permanent `url` and `posterUrl`: clips are public and stay reachable\nuntil the player deletes them.",
        "tags": ["media"]
      }
    },
    "/api/v1/media/{id}": {
      "patch": {
        "operationId": "media_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMediaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The uploader corrects their own clip's title, description or rating.",
        "tags": ["media"]
      },
      "delete": {
        "operationId": "media_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["media"]
      }
    },
    "/api/v1/media/{id}/rating": {
      "patch": {
        "operationId": "media_rate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RateMediaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Removes one of your own clips. The previous one in that category becomes\n the current claim again.",
        "tags": ["media"]
      }
    },
    "/api/v1/media/{id}/rating/history": {
      "get": {
        "operationId": "media_ratingHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "What that rating was before each change.",
        "tags": ["media"]
      }
    },
    "/api/v1/media/{id}/like": {
      "post": {
        "operationId": "media_like",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["media"]
      },
      "delete": {
        "operationId": "media_unlike",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["media"]
      }
    },
    "/api/v1/media/{id}/view": {
      "post": {
        "operationId": "media_recordView",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["media"]
      }
    },
    "/api/v1/media/{id}/engagement": {
      "get": {
        "operationId": "media_getEngagement",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Counts, plus `likedByMe` when a token is present — one like per account.",
        "tags": ["media"]
      }
    },
    "/api/v1/media/{id}/comments": {
      "post": {
        "operationId": "media_comment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMediaCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["media"]
      },
      "get": {
        "operationId": "media_listComments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["media"]
      }
    },
    "/api/v1/media/comments/{commentId}": {
      "delete": {
        "operationId": "media_deleteComment",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["media"]
      }
    },
...
  },
  "info": {
    "title": "FotSpot API",
...
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "The `accessToken` returned by any `/auth/*` login endpoint."
      }
    },
    "schemas": {
...
      "MediaCategory": {
        "type": "string",
        "enum": [
          "PACE",
          "DRIBBLING",
          "PASSING",
          "FINISHING",
          "PHYSICAL",
          "TECHNIQUE",
          "MATCH_HIGHLIGHTS"
        ]
      },
      "RequestUploadDto": {
        "type": "object",
        "properties": {
          "category": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MediaCategory"
              }
            ]
          },
          "filename": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["IMAGE", "VIDEO"]
          },
          "contentType": {
            "type": "string",
            "description": "Sent to R2 so the object is served back with the right type.",
            "maxLength": 100
          }
        },
        "required": ["category", "filename", "type"]
      },
      "ConfirmUploadDto": {
        "type": "object",
        "properties": {
          "category": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MediaCategory"
              }
            ]
          },
          "rating": {
            "type": "number",
            "description": "The player's own 0–100 claim for this attribute, evidenced by the clip.\n\nRequired for the six attribute categories and rejected for\nMATCH_HIGHLIGHTS — enforced in the service, because \"required unless the\nvalue of another field is X\" is not something class-validator states\nclearly enough to be worth the custom constraint.",
            "minimum": 0,
            "maximum": 100
          },
          "storageKey": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["IMAGE", "VIDEO"]
          },
          "title": {
            "type": "string",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 1000
          },
          "posterKey": {
            "type": "string",
            "description": "Key of the cover frame, from the same upload ticket. Optional: capture can\n fail, and a clip without a cover beats a refused upload.",
            "maxLength": 512
          }
        },
        "required": ["category", "storageKey", "type"]
      },
      "UpdateMediaDto": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "title": {
            "type": "string",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 1000
          }
        }
      },
      "RateMediaDto": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          }
        }
      },
      "CreateMediaCommentDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          }
        },
        "required": ["body"]
      },
...
    }
  },
...
}